+++ core/src/main/java/com/sonicgdx/sonicswirl/SensorReturn.java
@@ -2,9 +2,9 @@ package com.sonicgdx.sonicswirl;
 
 public class SensorReturn {
     public final Tile returnTile;
-    public final int returnDistance;
+    public final float returnDistance;
 
-    public SensorReturn(Tile returnTile, int returnDistance) {
+    public SensorReturn(Tile returnTile, float returnDistance) {
         this.returnTile = returnTile;
         this.returnDistance = returnDistance;
     }
